requirements.txt
click==8.0.3
Pillow==8.4.0
